backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
scikit-learn==1.4.0
numpy==1.26.3
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
joblib==1.3.2
tqdm==4.66.2
matplotlib==3.8.3
google-generativeai==0.3.2
python-dotenv==1.1.0 